backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
numpy==1.26.4
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9
